+++ CHANGELOG.md
@@ -13,7 +13,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
 - **breaking:** core: fix buffer overflow in function utf8_next_char and return NULL for empty string
 - **breaking:** core: fix integer overflow and return "unsigned long" in function util_version_number
 - core: write configuration files on disk only if there are changes ([#2250](https://github.com/weechat/weechat/issues/2250))
-- core: always enable partial completion for templates in option weechat.completion.partial_completion_templates ([#2253](https://github.com/weechat/weechat/issues/2253))
+- core: always enable partial completion for templates in option weechat.completion.partial_completion_templates, add option weechat.completion.partial_completion_auto_expand to expand word on new completion ([#2253](https://github.com/weechat/weechat/issues/2253))
 - core: add script name in output of `/debug hooks <plugin>`
 
 ### Added